refactor(Info): change casing
Since now, a field / property that starts with a capital letter *can* and
is supposed to be changed

If it starts with a lowercase letter, it's readonly. Kinda like private
since in most situations it won't be useful outside of the class, but
can still be read if there's any use for doing so.

Changing lowercase starting fields / properties is supposed to break the
program