feat(info->destroy): add hotkeys to kill numbered infos
The windows key + a number (up to 9, inclusive), destroys the info in
that index.

Say you have 10 infos visible currently.
You can press windows + 5 to destroy the fifth info.
If you press that hotkey again, it will do the
usual windows function (activate some app).
You also don't have a way to destroy the 10th info in this case,
aside from clicking on it with the mouse