make ValidateKey public
You can now pass a string to the ValidateKey method, that string will be
considered the valid characters. If you don't specify it, it will
consider Registers.ValidRegisters as the valid character string

The reason for this is because I use the Register class' validating
system in multiple places in my script and I would like to be able to
define a broader spectre of possible characters for something that
doesn't involve writing a file with that character in its name.

Technically doing this entire thing is pretty bad, because I'm using
something in a way it wasn't supposed to be used, creating dependencies
on it as if it's supposed to. So you, the person reading this, don't do
this in your project! In theory, ValidateKey should be moved outside of
registers, but I don't know if I'm gonna be bothered to