implement Map.Prototype.Choose(keyName)
Linker is a restricted idea, even though it was all I was going to use
it for

Now map objects have a Choose method, that lets you pick a value by
picking a key from a list of keys in Infos, that match the search
provided as a parameter

In other words, imagine your map is this:

```ahk
mappy := Map(
    "error objects",
    "https://youtu.be/JUyAq4t5XSQ",

    "how to make text objects",
    "https://youtu.be/wj7U-DNALN0",

    "text objects",
    "https://youtu.be/nU4vRFMf4F8",
)
```

Now when you do this:

```ahk
mappy.Choose("object")
```

You will see four infos: first one with nothing, to make you able to
cancel the choosing.

The other three are the key names.
All of the keys in the example match, so that's why you see all of them.

You pick an option by destroying its Info (clicking, for example)

Once you choose, that key's value gets returned

This whole idea is nice for situations where you dynamically get data
from maps and don't remember the actual key name, so want to search for
it