Info: use Function keys for specific destruction
the F1-F12 keys are pretty rarely used by default, enough so to be a
good set of hotkeys to destroy specific infos. Generally, pressing
windows + numbers is non-viable up to the key 6, since pressing windows
and a number before six is pretty janky. I was only able to do so
because I remapped a key on the right to be the windows key