HoverScreenshot: implement UseRecentScreenshot
Sets the last taken screenshot (by windows + shift + s) to be the
picture selected
Couldn't use a normal loop files, because A_LoopFileTimeModified is
imprecise.
When you take a screenshot, two pictures are written to the folder,
one of them being the thumbnail, pretty much immediately.
So closely, that using A_LoopFileTimeModified doesn't help differentiate
between them
However, the DllCall made in GetFilesSortedByDate *can* differentiate
between the two, and actually see which picture was latest.
And from there, you just pick the second one, which is *not* the
thumbnail, but is instead the latest screenshot.
Done, finally! Pretty much all thanks to @anonymous1184's help