WindowManager.Close waits at least a second for the window to close
The reason for doing this is the recent bug I've been getting where
clicking once with my mouse hotkey that runs this method actually ends
up closing two windows in a row.
Not sure why that's the case, but waiting at a second / until the window
is actually closed should solve the issue.
If that doesn't, I might just add a sleep