XButton scrolls switched around
The volume change on xbutton2 is something that's being activated many
times per second, because of which some xbutton2's slipped out from time
to time.
Undo / redo are actions you usually don't spam as much, giving you less
chances to slip up into a different action.
So now volume changes are on xbutton1, and undo / redo are on xbutton2