REMOVE LINKER
Managing multiple link files is actually annoying.
If I don't know the name of the link pretty accuretely, or at least a
part of it, I now need to go into the file to search for it.
It's not like this happens often at all, but still, there are a lot of
links I simply forget about, which is non-ideal.
I'd like to have a fuzzy search for links, and you know what will get me
exactly that? Storing them all in a json file!
Yes, it's that simple.
So now all my links are private and are stored in a json file.